androidStartupMetric: re-introduce First Contentful Paint
The previous failures were caused by enabling Perfetto, which since has
been reverted.

Re-introducing the metric with a fallback to scanning all processes,
just like for 'messageloop_start_time' - this should make the metric
resilient to missed process names. The issues with missing process name
remain for traces that are combined with systrace.

Bug: chromium:886621
Bug: chromium:893514
Change-Id: Ib1c2ca5b426a67c7071209ae7969fef4d5651d08