Reduce verbosity of typ output for catapult tests.
Verbosity was increased back in 2016 to help debug a telemetry issue, which has
since been fixed.

Increased verbosity makes it difficult for non-experts to read the log to
understand or even find the source of a test failure. This CL reduces the
verbosity back to standard levels, which only emits output for failed test runs.

Bug: chromium:891898, chromium:457446
Change-Id: I3bf0b115eeca011148ae093f287ead3edaeb6c9c
notry = true