Cache session ids in v2spa service worker.
v2spa stores session state in the URL. When the session state is too complex
to fit in the URL, it posts the state to the backend, which stores it and
returns an identifier for the state that the frontend can store in the URL.

SessionIdCacheRequest allows v2spa to post session state as often as it wants
without posting the same data to the backend more than once.
SessionIdCacheRequest uses the same algorithm as the backend to generate the
session id, which allows the frontend to display session ids without waiting
for the round-trip with the backend.

Bug: catapult:#4461
Change-Id: I82259a339693207051fd7e049ccc4db2211c7a55