[tracing] Restore the minimum mutator utilization metric.
This adds the "v8-gc-mark-compactor-mmu-100ms_window" metric.

The metric was introduced initially in [1] and later removed
in [2] in favor of the expected queuing time metric. In the
meantime we found that the metric is still useful for evaluating
GC optimizations and have been running it locally.

This patch restores a simpler version of the metric:
- only mark-compactor related events are used for computation.
- only the minimum mutator utilization is reported.

[1]: https://codereview.chromium.org/1908513002
[2]: https://codereview.chromium.org/2769653002

Change-Id: I97e82f2f9efaa801b4ac732a86bbc9dedd9dc27e