event_finder_util: skip renderer events when no mainThread
In http://crbug.com/891797#11 there is a link to an example trace that
complains with: "Import Warning: UserModelBuilder: TypeError: Cannot
read property 'sliceGroup' of undefined".

With this change the error goes away. Not sure how to test it properly
though, but I'm OK landing this oneliner as is without tests.

Bug: chromium:891797
Change-Id: I7046c48a8251b1378e8bb419deb41362c932a52a