Import perfetto json traces
Small changes to allow perfetto traces (in json format) to be imported
correctly. The previous implementation was ignoring the android process
dumps, this is now fixed. Also previously we assumed that every event
would have a thread name so the first thread name was saved and never
overwritten. Now we want to overwrite the thread name if was originally
unknown ("<...>").

Bug: catapult:#
Change-Id: I29926f6b202dcae9b82fb5c3b1eff9f788ba00c0