Update patch set 2
Patch Set 2:

(2 comments)

I have a few comments that we'll need to address in this patch.

In addition, we don't need to address this in this patch, but it'd be good to be a little more formal about what we want the interaction between all of the command line args to be, to be explicit about how --repeat and --retry-limit interact, etc. Ideally we can specify all of that as part of bit.ly/chromium-test-step-api, which we should revamp and bring up to date as part of this work.

Patch-set: 2
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>