[sounwdave] Revert to old /describe dashboard API
A recent change meant the API should now be called as:

  /describe?test_suite={test_suite}

However it seems that the change is not yet deployed (dashboard responds
with 404), but still works fine with the older interface:

  /describe/{test_suite}

So reverting back to using that.

This change can be reverted after the new API interface is confirmed
to be live.

TBR=benjhayden@chromium.org

Bug: catapult:#4461
Change-Id: Ie4efca3a3aa5aca1b18063ec1010fe018e7304bb