Add RecommendedOptions to v2spa
This is little more than an additional option-group that allows a few options to
be offered in two places: once in the full list of options, and again in a
shorter list. Selecting an option from either group adds/removes it from the
same selectedOptions array.

Demo: https://v2spa-dot-chromeperf.appspot.com/
See menu-input in alerts-section.

Bug: catapult:#4461
Change-Id: I7856206f160c5f766eac2d81be68c2ada95b66ef