Flip the default for ensure_logs_on_timeout
This will reduce the timeout by 5% for all tests, but will ensure that stdout is
returned on timeout making it easier to debug timeout issues in tests.

Note that this does not change the default for internal AdbWrapper functions as
those are also used by other functions than RunShellCommand. I will flip the
default for them in a later CL to avoid breaking too many things at once :-).

R=jbudorick@chromium.org, perezju@chromium.org

Bug: chromium:899721
Change-Id: I747e8300d673e95604a8ea1376308d9ba91e8e81