[telemetry] Fix bad timeout=None in platform.PathExists
Passing timeout=None is triggering a bug in the logic of
RunShellCommand.

The intended behaviour was to use the default values for each of
timeout/retry when they are not supplied. To easily support this,
instead we just pass all nameless **kwargs from platform.PathExists
to the lower level device.PathExists.

TBR=nednguyen@google.com
Bug: chromium:900877
Change-Id: If22cd315674bf44f1582727df90dd063217b3034