Copy existing expectations parser into typ.
We want to add support for handling the bit.ly/chromium-test-list-format
for test expectations to TYP.

This CL is the first step in that process; it simply copies over the
existing parser from py_utils unchanged. Subsequent CLs will update
the parser to handle the correct format, clean up the coding style,
and integrate it into TYP itself.

Bug: chromium:835690
Change-Id: I6d61975d34db868a6b551c001ad26e45928cf671