Update the expectation parser in TYP to match the new format.
The "tagged list" file format in bit.ly/chromium-test-list-format
drifted a bit from what was implemented in common/py_utils. Most
notably, we switched to using "[" and "]" to enclose the list of
tags, so that the list of tags could span multiple lines.

This CL also updates the json_results.py class to be consistent
with the spec by using "PASS" instead of "Pass", etc.

Bug: chromium:835690
Change-Id: I533d801b05d61d4f3c8dc15ef9c3f22b10308f6f