androidStartupMetric: workarund for missing main entry point marker
The assert in the bug fires because the code assumes FCP and messageloop
start events come in pairs. This assumption seems to be violated
currently with Perfetto ON when running on low end devices - it takes
some time to start the tracing machinery, so we are missing early
events.

As a workaround, ignore the initial FCP event in this case.

Bug: chromium:900878
Change-Id: I6bc67ff747deec6fb5c24dca441f526ac26da524