[telemetry] Treat most exceptions as "possibly handleable"
Currently Telemetry treats as "handleable" errors---which cause a story
run to fail, but allow the rest of the stories to continue---a few
whitelisted ones and all own (telemetry) exceptions.Error's; while
treating most other errors and "unhandleable"---causing the whole
benchmark execution to halt.

This is often too drastic, as many errors raised by Telemetry clients
are in fact "handleable" and should not stop the execution of the
entire benchmark due to an error in one of its stories.

In this CL we define instead a small set of explictly "unhandleable"
errors (KeyboardInterrupt, SystemExit, etc) which should definitely
halth the benchmark excecution, while most other Exception's are
treated as "possibly handleable", allowing other stories to continue
running but restarting the shared state to be safe.

Bug: chromium:900909
Change-Id: I5f23d5c33319b04f7631f0f3e77fe2a2bb62c131