Update computation of load bias to handle CrOS edge cases.
We've now seen examples on CrOS where an ELF has several segments, and some of
the segments do not report the proper name in the memory map heap dump. This CL
updates the load bias computation to assume that all maps with the same name are
from the same ELF, and uses the start_address of the first such map to compute
the load bias of subsequent maps.

Bug: chromium:904492
Change-Id: I33344afdec7107a8c8f69e94cff4aec000154519