Fix merging single-bin Histograms with multi-bin Histograms.
Single-bin histograms' binBoundariesDict_ is undefined.
Both canAddHistograms and addHistograms fail when trying to merge single-bin
Histograms with multi-bin Histograms.
This CL fixes these methods to behave reasonably in this case.
If so many samples have been added to the single-bin Histogram that it has
dropped some sample values, then bin counts for merged multi-bin Histograms will
be inaccurate, so percentiles may be inaccurate, but RunningStatistics will be
correct. Metrics can set maxNumSampleValues to reduce the likelihood of this
situation, at the cost of storing the sample values.
It is still recommended for metrics to avoid using single-bin Histograms by
either allowing Histogram to use the per-unit default bin boundaries, or
explicitly setting sensible bin boundaries.

Change-Id: Iee67210f711b6a2341bd5dae1a9c0796638b406b