Add a way to visualize when raster tasks are performed.
This CL adds a graph to represent visually when the raster tasks are
performed in comparison to other tasks in the graphics pipeline. Metrics
Visualization was also refactored to organize the UI elements better,
to avoid putting all the code for this in one file, and so that the
histograms wouldn't have to be processed more than once.

Bug: catapult:#4562

Change-Id: Ia6bdf19587052d36016101b6f60d21770b38ebbd