Update patch set 1
Patch Set 1:

Maybe it'd be better to commit this on Monday? People in Canada and UK may still be working/committing CLs to catapult, and it may be hard for them to get support over the holiday if this breaks.

Patch-set: 1
Reviewer: Gerrit User 1001913 <1001913@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>