Move typ's expectation-processing code into expectations_parser.py.
Prior to this CL, the logic for handling test expectations in typ
was mostly split between runner.py and expectations_parser.py.
runner.py was a fine location for actually loading the file contents,
but most of the logic of actually parsing and handling the expectations
isn't all that germane to running the tests, so it was bloating the
code. Plus, having a file dedicated only to parsing the tagged test list
format was a bit excessive.

So, this CL moves the logic for figuring out the expected results for
a given test into expectations_parser.py and renames some of the classes
a bit to be more generic. This simplifies runner.py a bit and complicates
expectations_parser.py, but that's probably the right tradeoff.

It also makes expectations_parser.py a less-accurate name for the file.
It should probably be renamed to expectations.py, but I avoid doing that
in this CL to minimize the size of the diff to review.

Bug: chromium:835690
Change-Id: I4c8cc16547489c409d9cab6fa7b10910582ad83e