Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> One of the issues with this approach is that I can't print the list for android devices (e.g. rendering.mobile) without having an android-device connected. Is there a way to get past that?

what happens if you try it. i.e. set --browser=x --print-only. where is the error? can we just plumb through the print-only flag to ignore the error in that case?

The other way is to make 
./run_benchmark list
be similar to 
./run_benchmark run

s.t. for most 
./run_benchmark run --blah=blah ...
you can replace with
./run_benchmark list --blah=blah ...
and it just works.

Otherwise, if you don't have time to do that then I'm okay with your earlier approach of adding it to list, with the caveat that you should try to make the filter feel similar to waht we already have.

Patch-set: 1
Reviewer: Gerrit User 1149061 <1149061@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>