Add TimeseriesCacheRequest to v2spa service worker.
This allows the service worker to cache timeseries data. The two major
sources of complexity are the revision range and the columns. Every
timeseries is effectively a 2D dataframe that is too large to always
fetch completely, so v2spa fetches just the sub-areas of the table that
it needs. It may fetch sub-areas that overlap with previously-fetched
sub-areas in either or both dimensions. The page may also issue multiple
overlapping fetches simultaneously, requiring this class to deduplicate
in-progress requests.

This is mostly from sbalana.

Bug: catapult:#4461
Change-Id: I8b62308a41a77d1864796f159e84a4fc764b4a86