Update patch set 13
Patch Set 13:

Is visualization-data-container a better name? It functions as a parent for the graphs generated when the rendering benchmark is run and it extracts the data from the histograms. I build the data for each chart in this file because it seemed easier considering the way the data from the histograms is stored. If you have any suggestions for improvement let me know.
Thanks, 
Linwan

Patch-set: 13
Reviewer: Gerrit User 1305608 <1305608@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>