Adds new verification to expectation_parser.py
Currently typ accepts tags that are a part of multiple tag sets.
This change will add a new verification that will make sure each
tag set is mutually exclusive. If this rule is violated, then a
ParseError exception will be thrown and will contain an appropriate
message. I also added a new unittest,
testParseExpectationMutuallyExclusiveTestSets in
expectations_parser_test.py for this change

R=dpranke@google.com, nednguyen@google.com

Bug: chromium:905426
Change-Id: Ie63cfa4957578e289e694a2f582008e7885c5eb2