Remove vendor prefixes from expectations device type names
I got those device type names from the Geekbench model names, but it
looks like whatever Telemetry does to populate the device type names
strips the vendor prefix that's present on Geekbench. The end effect of
this was that the device type name query wasn't returning true and we
were failing to disable on the appropriate WebView devices.

Bug: chromium:911181
Change-Id: I4402de002d30bab15be9ad72195d22bc7e90945d