Add a way to visualize when raster tasks are performed.
This CL adds a graph to represent visually when the raster tasks are
performed in comparison to other tasks in the graphics pipeline. Metrics
Visualization was also refactored to organize the UI elements better,
to avoid putting all the code for this in one file, and so that the
histograms wouldn't have to be processed more than once.

The document explaining this feature will be updated and a new
results_demo.html file will be uploaded once chromium:902009
is fixed.

The document: https://docs.google.com/document/d/1nUiPkFEinH9hYxJsw_JVzXXaYkwdYxEB9yFjqEwmVzw/edit?usp=sharing

Bug: catapult:#4562

Change-Id: Ia6bdf19587052d36016101b6f60d21770b38ebbd