Update patch set 3
Patch Set 3: Code-Review+1

LGTM once the bots failures are fixed.

Do you know if anyone is working on eliminating the source of floating point errors?
Or is it fundamental problem? It seems that the start and the end of a task can be precise (in theory)

Patch-set: 3
Reviewer: Gerrit User 1002405 <1002405@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1