android_browser_finder: install extra apk
In order to install MapsWebApk.apk in http://crrev.com/c/1352315 add an
API to PossibleAndroidBrowser to install APKs from "$CHROMIUM_OUTPUT_DIR/apks".

Initially I thought it could be made during browser_finder.FindBrowser()
by extending the |finder_options|, but it did not work because the
platform was not initialized at that point.

Bug: chromium:760498
Change-Id: Ide5e54a21ffad2dcdcf41eb99ab458014cef5e5c