Update patch set 5
Patch Set 5: Code-Review+1

lgtm but can you create a test for this? I think the test can be in page_run_end_to_end_unittest

I imagine the test would be creating two pages. The first page scroll to the end of the page. Then in the 2nd page, you add an assertion that the page is started with the scroll view at the top of the page. In the test case, you would also use a NoClosingBrowserSharedState which inherits from SharedPageState and overrides ShouldStopBrowserAfterStoryRun to return False.

Patch-set: 5
Reviewer: Gerrit User 1128435 <1128435@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1