Fetch list of sheriffs in v2spa
The V1 alerts page receives the list of sheriffs every time it requests alerts.
There is no way to just get the list of sheriffs.

This CL adds a way for v2spa to fetch the list of sheriffs separately from
alerts. They are cached in the service worker similarly to test suite names and
report template infos.

Bug: catapult:#4461
Change-Id: I3f3ccd723d264ceab599a1894b155e28c4c8ab9b