chromeos: close all tabs before running the next test
On ChromeOS we don't stop the browser between each story. Further, we keep one
tab open. This means if the next story loads the same url as the previous story
the next story gets whatever state the previous story ended in. For example,
some scrolling tests load the same page and scroll to the end of the document.
Without the change here, only the first story did anything, all the following
stories completed immediately because the page was already scrolled to the end.

BUG=chromium:911777
TEST=none

Change-Id: I34800fa9ec239464b4d85fed7ec0f0ffd353c447