Add AndroidPlatform.InstallApk()
In order to install MapsWebApk.apk in http://crrev.com/c/1352315 add an
API to AndroidPlatform to install APKs from "$CHROMIUM_OUTPUT_DIR/apks".

Initially I thought it could be made during browser_finder.FindBrowser()
by extending the |finder_options|, but it did not work because the
platform was not initialized at that point.

To avoid duplication in android_browser_backend_settings move the code
to find the APK to a common place: telemetry.core.util.

It seems that these functions would not produce anything useful with
chrome_root being not equal to util.GetChromiumSrcDir(). Remove plumbing
of this parameter to simplify.

Bug: chromium:760498
Change-Id: Ide5e54a21ffad2dcdcf41eb99ab458014cef5e5c