Add telemetry.core.util.FindLatestApkOnHost()
In order to install MapsWebApk.apk in http://crrev.com/c/1352315 add an
API to install APKs from "$CHROMIUM_OUTPUT_DIR/apks".

Initially I thought it could be made during browser_finder.FindBrowser()
by extending the |finder_options|, but it did not work because the
platform was not initialized at that point.

Since --chrome-root needs to be respected, the shared state will pull it
from |finder_options| and push down to FindLatestApkOnHost().

To avoid duplication the android_browser_backend_settings is using the
same function.

Bug: chromium:760498
Change-Id: Ide5e54a21ffad2dcdcf41eb99ab458014cef5e5c