Split metrics_visualization into two files
This cl splits metrics-visualization into 2 classes in preparation
for more charts that present the data in other ways. Previously the
metrics-visualization object contained and managed all the charts
created for different sets of metrics in one object. This was workable
when the charts were all similar, but it needed to be cleaner to
support more features. I wanted something that would only process
the histograms once and use that data for all the charts, while
simultaneously allowing each data set to have their own object that
specifically manages the charts for that data set.

The new object, visualizations-data-container, processes the histogram
inputs for use in all types of charts around the different types of
graphs. It’s also a container for all of the data set visualization
objects. The data passed along to each chart is processed in this object
because of the way the processed data was organized.

The new metrics-visualization manages the stacked charts for one set of
metrics. This set of metrics is defined as an entry in METRICS. It has
all the same features as the previous metrics-visualization, except it
does not process its own histograms, and it only manages one set of charts
at a time.

The cl sets up for my next cl which will add a new type of chart to the
tool.

Bug: catapult:#4562
Change-Id: I36b357c3cd84ab504a52767fd78b31fe436310ba