Changed average raster task times metric
Previously the raster start times were calculated based on the
average start times of all the slices connected to a frame.
Now it takes the start value of the first raster task instead.
A metric was added to find the end value of the last raster task
for a frame to help calculate the total duration of raster tasks
for a frame.
The way the CompositorTileWorker threads were found was also
changed because a function used previously was deprecated.
Unit tests were also updated to reflect these changes.

chromium:911190
Change-Id: I158b4a387c8bc6dac8c4b76d6cc0e7032b774659