Update patch set 2
Patch Set 2: Commit-Queue+1

(2 comments)

> Patch Set 1:
> 
> (2 comments)
> 
> Is there a plan for updating the sizes script to produce Histograms directly? Would that be easier than adding this temporary conversion script?

That's the eventual goal, but it's not necessarily easier to do that since the gtest perf tests don't currently rely on Catapult, so we'd have to make sure that's added as a dependency everywhere then change the test runner logic. This way, the tests can continue running as they always have and we can do the conversion on the builder where we know that we have Catapult available. Since this is potentially blocking LUCI migrations which need to be done by March, I've prefer to get this simpler working solution in then cleanup/remove usage of conversion scripts at a later time.

Also note that this is for gtest perf tests, which are different than the sizes script.

Patch-set: 2
Reviewer: Gerrit User 1163549 <1163549@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1