[pinpoint] Parse isolate hashes from buildbucket.
We're encountering what I think is a race condition, where the build
is complete but the isolate hashes are not yet in the Datastore.

Buildbucket already has the isolate hashes in the build properties,
so we can just parse them out when we poll for the build status.

Depends on https://crrev.com/c/1366977

Bug: chromium:868926
Change-Id: Ieff11da033f4c582d6e594855a8605e741f21f91