Android: pull more lines from logcat by default
I am try-running http://crrev.com/c/1352315 via Pinpoint using the bot
android-go-perf. Example job:
    https://pinpoint-dot-chromeperf.appspot.com/job/14eb5e40940000

Individual tasks reveal in the logs that browser crashes appeared, like
this:
  - android:id/alertTitle['Clankium has stopped']

The fragment of the logcat that Telemetry attaches does not contain the
information about the crash, which complicates the investigation a lot.
(!)

The reason for the crash to be missing: only 500 lines are being pulled
by Telemetry (running command "logcat -d -t 500") which covers about 30
seconds of idle time. During these last 30 seconds Telemetry had been
waiting for devtools connection, and the crash causing he connection to
be refused happened 40-50 seconds earlier.

Increasing the limit from roughly 30 seconds to 90 seconds should give
enough room. This translates to 3x the number of lines -> 1500.

There are probably greener ways to do it, but I'd prefer this one
character fix in the short term to unblock the investigation.

Bug: chromium:760498
Change-Id: Ifd423c650082fe0f886d6c211520f0a48eaa0121