Reland the safe parts of "Syntax updates for Polymer 2.0."
Polymer 2.0 is production ready, and we advise all projects to begin
migration. To aid in that migration, we're writing and applying some
automated fixes for the mechanical parts of the upgrade.

The <content> element is deprecated, part of the Shadow DOM v0 API which
has been supplanted by Shadow DOM v1's <slot> element. For very simple
uses – like those in this PR – this change should be safe and will not
require changes to the usage sites of your element. Therefor, these
changes are backwards-compatible with Polymer 1.0 applications.

As <content> is the more powerful API, Polymer 1.0 supports both
<content> and <slot> by converting <slot> back into <content> at
runtime. Polymer 2.0 only supports <slot>. There will likely never be
cross-browser native support for <content>, while Safari and Chrome
already support <slot> natively.

This change also updates usage of `@apply` to remove the parentheses.
Like the other changes in this CL, these are backwards compatible with
Polymer 1.0, and necessary for Polymer 2.0.

Original pull request: https://github.com/catapult-project/catapult/pull/4016
Original gerrit CL: https://chromium-review.googlesource.com/751690
Revert: https://chromium-review.googlesource.com/769447

This reland is a more direct copy of the original pull request, 4016.
The original gerrit CL, 751690, incorrectly migrated some content
elements that are not contained in dom-modules, such as the content
element in overlay.html.
Using slot elements in v0 shadow roots (such as created by
createShadowRoot in overlay.html), outside of Polymer dom-modules is not
supported.
This CL keeps the content element in overlay.html.
When Overlay and RecordSelectionDialog are upgraded to Polymer
dom-modules, their content elements should be changed to slot elements
at that time.

We can keep using ::content CSS rules with slot elements in Polymer 1.0
because Polymer 1.0 converts the <slot> elements to <content> at runtime.
However, Polymer 2.0 does not, so when trace viewer migrates to Polymer 2.0,
the ::content CSS rules will need to change to `::slotted`. However, there
are some caveats. ::slotted isn't as powerful as ::content, so we may need
to redesign some elements to not rely on ::slotted.

Bug: chromium:784826
Change-Id: I0fb3ca680c83c53297411d27fbfc4aa4fc2e2961