[py_utils] Add modules_util.RequireVersion
The core of Telemetry's external_modules.ImportRequiredModule is
extracted to a py_utils module to allow reuse by other catapult
projects.

This is in preparation for the eventual deprecation of
external_modules itself, which is due to be replaced by the use
of vpython dependencies instead.

This CL also uses the new module in devil's device_recovery.py
which already runs using vpython everywhere.

Bug: chromium:776709
Bug: chromium:777865
Change-Id: I68ecae5c67c821b8aa809507eac826f0457af5c5