Use storyTags instead of tagMap in update_test_suite_descriptors.
This is simpler and will allow removing the TagMap diagnostic entirely, though
it requires 1 SparseDiagnostic query per test case instead of per bot. There are
only a few hundred test cases, and the queries are parallelized, so this
increase is negligible.
A followup CL will remove TagMap.

This chart monitors a logs-based metric for the mean wall-time to process a single
TestMetadata key. Collecting tags is aggregated into this as overhead.
https://app.google.stackdriver.com/dashboards/16092431051342960751?project=chromeperf

Bug: chromium:918218
Change-Id: I55c00058b0177461262d0d9126ab17007a04adc9