Removes expectations_file named argument from constructor of ProjectConfig
This CL is part of the effort to rename the expectations_file argument
of the telemetry.project_config.ProjectConfig constructor. In a previous
CL we removed the use of the expectations_file argument in
tools.perf.chrome_telemetry_build.ChromiumConfig constructor. In this
CL we will remove the argument from the ProjectConfig's classes constructor.

R=nednguyen@google.com

Bug: chromium:835690
Change-Id: I42a454228c8574f8a7700cc2e576b524727a7a23