Fix request-base-test with polyfilled html imports.
Native HTML Imports guarantees that imports are loaded before the 'load' event
fires, but that is not the case with polyfilled imports, so suite()s cannot rely
on imports. test()s can rely on imports, however. v2spa's request-base-test
fails with this flag because it defines a subclass of cp.RequestBase at suite()
call time.

This CL fixes the test by waiting to define TestRequest until test() call time.

Change-Id: Ic0408b15866886c06e4c32513e57b35d9ad66332