Relax an assertion in heap_profiler_unittest.
This test asserts that the heap profiler measured the sizes of 12 properties of
a Histogram object.
This assertion has changed a few times recently as I've been changing the
Histogram class, but it might also not be as stable across platforms as I'd thought:
https://chromium-review.googlesource.com/c/catapult/+/1227065

This CL relaxes the assertion to check that at least a few properties were
measured in order to make it easier to change the Histogram class and more
robust across platforms.

Change-Id: Iee3915cabba0ec9d055af42c085275f83e1b32b8