Remove support for legacy json formats from render_histograms_viewer.
This module renders results.html files with HistogramSet JSON.
If the results.html target file already exists, it reads existing results in
legacy formats that haven't been used in a while. The oldest supported legacy
format is 2 years old. The newest is 3 months, which is two full chrome release
cycles.

Change-Id: I7adaa648f11286d35805a19dbe9bfd6ae4d415e1