Remove TagMap.
Telemetry benchmarks can contain hundreds of test cases, so they tag them to
make it easier to filter them. The tags are stored in the storyTags GenericSet
diagnostic, which is used by results.html, and soon by the dashboard:
https://chromium-review.googlesource.com/c/catapult/+/1394567
Story tags are also stored in TagMap diagnostics, which map from tags to sets of
stories. TagMaps contain the same information as storyTags, just centralized in
a different data structure.
After that change to update_test_suite_descriptors, TagMaps will not be used
anywhere. This CL removes TagMaps entirely.

Bug: chromium:918218
Change-Id: Icc8e637db54a56c03f0010baf9d8b95ed50ad195