[tracing] Ignore unrecognized diagnostic types
After the recent deletion of the TagMaps diagnostic, trying to use
histogram_set.ImportDicts on a histogram set that was created before
the diagnostic got removed (e.g. from pinpoint jobs) caused a
fatal error.

This CL downgrades the error to a warning, so that the method can
still be used with data coming from older histogram json files.

Bug: chromium:918218
Change-Id: I61718b146cf42c8d1e07ca536580b5010c0051e9