Reland "[py_utils] Add modules_util.RequireVersion"
The core of Telemetry's external_modules.ImportRequiredModule is
extracted to a py_utils module to allow reuse by other catapult
projects.

This is in preparation for the eventual deprecation of
external_modules itself, which is due to be replaced by the use
of vpython dependencies instead.

This is a reland of:
https://chromium-review.googlesource.com/c/catapult/+/1394564
without the changes in devil.

Bug: chromium:776709
Change-Id: Iad836eb24c9d0b0fb0f1c2d6b735ee20aee730c8
TBR=crouleau@chromium.org,sullivan@chromium.org