Fixes an issue with the JSON output of Typ
This issue happens when an expectations file is passed to Typ.
The JSON output for a skipped test is showing incorrect results.
This CL will introduce changes so that the expected result for a
test will show up, even when a test is skipped. It will also add
changes to make sure the is_unexpected flag and is_regression flag
does not show up when a test is skipped.

R=dpranke@chromium.org

Bug: catapult:000000
Change-Id: Ie149c1e5075f18a9e81d023ed1f9ffa7c9af3d25