Dashboard - Add helper to convert graphjson to histograms
This will be used by Pinpoint soon to run non-telemetry try jobs. We
can read the graphjson results in Pinpoint, but try jobs tend to want
to display a results2 table. IIRC the graphjson format isn't really
documented, but it's pretty straightforward. This just directly
converts to histograms + story diagnostics, and adds an option to add
a label diagnostic so that nice columns appear in the result.

Bug: chromium:908505
Change-Id: I46d25dfab8c1042ef615721cdbf77643240a3bc5