Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> (1 comment)
> 
> It looks like resampleMean is being called whenever histograms are deserialized, and again when merged. This seems unnecessary because the CI statistics may never be used. Please delay computing these statistics until the caller requests them by calling getStatisticScalar(). Feel free to memoize and clear the memoized values in addHistogram().
> 
> Would it make sense to condense the two statistics down to one by subtracting them?
> It takes more than twice as long to read two numbers than to read one due to mental load.
> 
> Please parameterize the CI statistic in the same way that the percentile and ipr statistics are parameterized. You may want 95% now, but you may need to tune that number after this hits the field.
> 
> Did you consider using an online algorithm like this one?
> http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.641.4627&rep=rep1&type=pdf
> I haven't finished reading this paper or the code in the linked github project, so I'm not sure that it would be appropriate, but I'd like to at least see a little discussion about online alternatives. If an online algorithm would be appropriate, then it would probably belong in RunningStatistics depending on how that might affect serialization.
> 
> Please add tests and translate the JS changes to the python version.

Thanks for the review,

I'll add the resampleMean to getStatisticScalar(), and also try to add parameters to CIs just like the percentile and ipr.

I don't think we can subtract the two statistics and only report the range, as the calculated mean does not necessarily in the middle of the range. For example the mean can be 12 when the confidence interval is 10 to 20.

About the online algorithm for bootstrapping, I was not able to understand the algorithm they were using in the paper, so I checked the research they based the algorithm on (https://ti.arc.nasa.gov/m/profile/oza/files/ozru01a.pdf) and apparently the algorithm they are proposing is about boosting, which is a supervised learning method, and not the bootstrapping I'm trying to implement.

Patch-set: 17
Reviewer: Gerrit User 1307714 <1307714@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>