Pinpoint - Support non-telemetry perf tests
The backend already supports running non-telemetry tests, but we just
need to make sure ones kicked off get the right isolate target and
that the graphjson reader is updated accept None arguments (like the
histogram reader).

There's a graphjson -> histogram converter supplied in tracing/value so
we just use that directly in results2.

Bug: chromium:904423
Change-Id: I721104fa0247f5c7df9c9bce5cf7ce270b933d57