Deserialize a new HistogramSet JSON format.
This will allow the dashboard and results.html to support a new format
alongside the existing format briefly while we refactor the
dashboard's handling of diagnostic guids. When the dashboard is ready,
another CL will switch everything over to the new format.

This is split out from 'Refactor shared diagnostics':
https://chromium-review.googlesource.com/c/catapult/+/1378593

Design doc:
https://docs.google.com/document/d/1agEZc6CmYRgMGY_IVjUc2b56Hsq5HZPbsPZ4n3KwLJc

Serializer:
https://chromium-review.googlesource.com/c/catapult/+/1392843

Bug: chromium:918208
Change-Id: I157e4174fd455da0b48ee99214c5e5a77c41935f