Update patch set 24
Patch Set 24:

> Patch Set 24:
> 
> (2 comments)
> 
> I can't really understand this changelist very easily. I recommend you to break it up into smaller changes or to try to explain it with some examples. 
> 
> For example, I am having trouble with this explanation:
> 
> > 2,This happens when an expectations file is passed. If a test calls self.skipTest and Skip is not in the expectations for that test in the file, then a regression is reported and a non-zero exit code is returned.
> 
> I don't understand what is wrong with this? It seems like that is how the test should work. But maybe I'm missing something. I need you to explain it a bit more.

I removed that part. I noticed an existing unit test that requires a regression to be reported if a test skipped unexpectedly by calling self.skipTest(). I will add some examples to the bug. The entire issue boils down to pretty much this. When an expectations file is given to Typ for a set a of tests, the expected field for each test in the JSON output should show the expected results specified in the expectations file. This CL will make changes so that it does.

Patch-set: 24
Reviewer: Gerrit User 1311766 <1311766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>