Devil: string constants for ABIs
No change in logic.

This adds named constants for Android's supported ABIs. This
intentionally omits mips (32 & 64) and armeabi because these are not
supported by the NDK.

These might be nice to make sure arch-specific infra checks the correct
ABI (reducing the possibility for typos).

This cross-links to the DeviceUtils methods which we can compare these
constants against.

This also replaces string literals with the new constant values.

Test: git cl presubmit
Change-Id: Iacc8c4d5763aac9f4e1f25b6bcf05c42a1bc4c08