Fix sum field in gtest_json_converter
Changes the sum field in gtest_json_converter to be count * mean instead
of just mean. This is because the mean of merged histograms is
calculated using the merged sums and merged counts. Without this change,
the data uploaded to the dashboard that was converted using the script
was halved.

Bug: chromium:921342
Change-Id: I353546659cf948792a1fc593bba14b14c09d5506