Fixing issue with tests skipped via the 'Skip' tag
This issue happens when an expectations file is passed to Typ.
When a test has the Skip tag in it its expectations in the file,
it is skipped. However the expected field in the JSON results shows
only PASS. This CL will make changes so that the expected field
will show all the expectations that were specified in the expectations
file.

R=crouleau@chromium.org, kbr@chromium.org

Bug: chromium:923494
Change-Id: I20fb56d91737363d957c4117093d2263ae7a0236