Reland: devil: string constants for ABIs
Partial reland of 713df15b1c90d3544325aae98bfeca820393099f
(http://crrev/c/1431852).

No change in logic.

This adds named constants for Android's supported ABIs. This
intentionally omits mips (32 & 64) and armeabi because these are not
supported by the NDK.

These might be nice to make sure arch-specific infra checks the correct
ABI (reducing the possibility for typos).

Unlike the initial CL, this does *not* use the constants in any other
devil files. This is postponed so that we can fix the pydeps files
downstream in chromium first.

Bug: chromium:924778
Test: Checked that pydeps files need not change yet
Change-Id: I5b7d51bae34adb12873ace08c5b5a755ce267a9a