Devil: use new constants for ABIs
Partial reland of 713df15b1c90d3544325aae98bfeca820393099f
(http://crrev/c/1431852).

This uses the new constants added in the other part of the reland. This
is so we can change pydeps files downstream in chromium.

This also fixes a bug which uses "x64" instead of the correct ABI value
("x86_64").

Bug: chromium:924778
Test: git cl presubmit (tests pass)
Change-Id: I947b6194eace3cd1f74a52b114413e1f5fbc3a59