Devil: use new constants for ABIs
Partial reland of 713df15b1c90d3544325aae98bfeca820393099f
(http://crrev/c/1431852).

This uses the new constants added in the other part of the reland. This
is so we can change pydeps files downstream in chromium.

This also fixes a bug which uses "x64" instead of the correct ABI value
("x86_64").

Bug: 924778
Test: git cl presubmit (tests pass)

Reland: devil: string constants for ABIs

No change in logic, this is a reland (with some reworks) of
713df15b1c90d3544325aae98bfeca820393099f (http://crrev/c/1431852).

This adds named constants for Android's supported ABIs. This
intentionally omits mips (32 & 64) and armeabi because these are not
supported by the NDK.

These might be nice to make sure arch-specific infra checks the correct
ABI (reducing the possibility for typos).

Unlike the initial CL, this does *not* use the constants in any other
devil files. This is postponed so that we can fix the pydeps files
downstream in chromium first.

Bug: 924778
Test: Checked that pydeps files need not change yet
Change-Id: I947b6194eace3cd1f74a52b114413e1f5fbc3a59